separate utils from main interface
